+++ src/main/resources/db/migration/V0.3.2__Create_challenge_completions_table.sql
@@ -3,7 +3,7 @@ CREATE TABLE `challenge_completions` (
     `challenge_id` INT NOT NULL,
     `team_id` INT NOT NULL,
     `completed_at` TIMESTAMP,
-    `seconds` FLOAT,
+    `seconds` DOUBLE,
     PRIMARY KEY (`id`),
     FOREIGN KEY (`challenge_id`) REFERENCES `challenges` (`id`) ON DELETE CASCADE,
     FOREIGN KEY (`team_id`) REFERENCES `teams` (`id`) ON DELETE CASCADE
